Replace half4 with float4

+++ sources/Metal/Shaders/iTermCursorRenderer.metal
@@ -64,12 +64,12 @@ iTermTextureCursorVertexShader(uint vertexID [[ vertex_id ]],
     return out;
 }
 
-fragment half4
+fragment float4
 iTermTextureCursorFragmentShader(iTermTextureCursorVertexFunctionOutput in [[stage_in]],
-                                   texture2d<half> texture [[ texture(iTermTextureIndexPrimary) ]]) {
+                                   texture2d<float> texture [[ texture(iTermTextureIndexPrimary) ]]) {
     constexpr sampler textureSampler (mag_filter::linear,
                                       min_filter::linear);
 
-    const half4 colorSample = texture.sample(textureSampler, in.textureCoordinate);
+    const float4 colorSample = texture.sample(textureSampler, in.textureCoordinate);
     return colorSample;
 }
